requirements.txt
alembic==1.17.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cs50==9.4.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
psycopg2==2.9.11
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.45
sqlparse==0.5.4
termcolor==3.2.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.4
wheel==0.45.1
